requirements.txt
django==3.0.2
django-filter==2.2.0
django-crispy-forms==1.8.1
django-widget-tweaks==1.4.5
pillow==7.1.2
django-sslserver==0.22
django_tables2==2.2.1
xmltodict==0.12.0
django-import-export==2.0.2
xhtml2pdf==0.2.4
requests
